requirements.txt
Flask==0.10.1
GChartWrapper==0.9
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.11.2
argparse==1.2.1
distribute==0.6.28
itsdangerous==0.24
matplotlib==1.5.x
mock==1.0.1
moin==1.9.7
nose==1.3.4
numpy==1.10.1
pyparsing==2.0.2
python-dateutil==2.4.2
pytz==2015.7
six==1.8.0
wsgiref==0.1.2
